Create command.hpp
Consider a "simple" switch. In this example we configure the Switch with two commands: to turn the light on and to turn the light off.

A benefit of this particular implementation of the command pattern is that the switch can be used with any device, not just a light. The Switch in the following C++ implementation turns a light on and off, but the Switch's constructor is able to accept any subclasses of Command for its storeAndExecute function parameter. For example, you could configure the Switch to start an engine.